+++ radio/er9x/src/er9x.h
@@ -1119,8 +1119,8 @@ void putsTime(uint8_t x,uint8_t y,int16_t tme,uint8_t att,uint8_t att2);
 
 #ifdef FRSKY
 void putsTelemetry(uint8_t x, uint8_t y, uint8_t val, uint8_t unit, uint8_t att);
-uint8_t putsTelemValue(uint8_t x, uint8_t y, uint8_t val, uint8_t channel, uint8_t att ) ;
-uint16_t scale_telem_value( uint8_t val, uint8_t channel, uint8_t *p_att ) ;
+uint8_t putsTelemValue(uint8_t x, uint8_t y, uint16_t val, uint8_t channel, uint8_t att ) ;
+uint16_t scale_telem_value( uint16_t val, uint8_t channel, uint8_t *p_att ) ;
 #endif
 
 extern int16_t calc100toRESX(int8_t x) ;
